// interfaces/dtos/order.dto.interface.ts
export enum orderStatusEnum {
  ACCEPTED = "ACCEPTED",
  CANCELLED_ORDER = "CANCELLED_ORDER",
  CANCELLED_PAYMENT = "CANCELLED_PAYMENT",
  CARD_PAYMENT_FAILED = "CARD_PAYMENT_FAILED",
  CONFIRMED = "CONFIRMED",
  DECLINED = "DECLINED",
  DELIVERED = "DELIVERED",
  DROPPED_AT_CHECKOUT = "DROPPED_AT_CHECKOUT",
  FAILED_PAYMENT_CONFIRMATION = "FAILED_PAYMENT_CONFIRMATION",
  FULFILLED = "FULFILLED",
  PENDING = "PENDING",
  PROCESSING = "PROCESSING",
  RETURNED = "RETURNED",
  SHIPPED = "SHIPPED",
  USSD_PAYMENT_FAILED = "USSD_PAYMENT_FAILED",
}

export interface IUpdateOrderStatusDTO {
  accessToken: string;
  additionalMessage: string;
  id: string;
  orderStatus: string;
  reason: string;
}